backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pymongo==4.7.2
python-dotenv==1.0.1
cloudinary==1.40.0
pinecone-client==3.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
sentence-transformers==3.0.1